+++ finagle-base-http/src/main/scala/com/twitter/finagle/http/FormElement.scala
@@ -19,5 +19,6 @@ final case class FileElement(
   name: String,
   content: Buf,
   contentType: Option[String] = None,
-  filename: Option[String] = None)
+  filename: Option[String] = None,
+  isText: Boolean = false)
     extends FormElement